refactor

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -9,9 +9,8 @@ import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repo.UserRepository;
 import ru.practicum.shareit.util.exeptions.ServiceException;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static ru.practicum.shareit.util.exeptions.ErrorMessage.REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID;
@@ -32,12 +31,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto getById(Long userId) {
-        try {
-            User user = userRepository.getById(userId);
-            return UserMapper.mapperUserToDto(user);
-        } catch (EntityNotFoundException exception) {
-            throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
-        }
+        Optional<User> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isEmpty()) throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
+        return UserMapper.mapperUserToDto(optionalUser.get());
     }
 
     @Override
@@ -55,27 +51,28 @@ public class UserServiceImpl implements UserService {
     public UserDto update(UserDto userDto) {
         User update = UserMapper.mapperUserDtoToUser(userDto);
 
-        try {
-            User user = userRepository.getById(update.getId());
+        Optional<User> optionalUser = userRepository.findById(update.getId());
 
-            if (update.getName() != null && !update.getName().equals(user.getName())) user.setName(update.getName());
-            if (update.getEmail() != null && !update.getEmail().equals(user.getEmail())) user.setEmail(update.getEmail());
+        if (optionalUser.isEmpty()) throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
+        User user = optionalUser.get();
+
+        if (update.getName() != null && !update.getName().equals(user.getName())) user.setName(update.getName());
+        if (update.getEmail() != null && !update.getEmail().equals(user.getEmail())) user.setEmail(update.getEmail());
+
+        try {
 
             user = userRepository.save(user);
             return UserMapper.mapperUserToDto(user);
+
         } catch (DataIntegrityViolationException exception) {
             throw new ServiceException(USER_ERROR__VALID_DUPLICATE__EMAIL);
-        } catch (EntityNotFoundException exception) {
-            throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
         }
     }
 
     @Override
     public void deleteById(Long userId) {
-        try {
-            userRepository.deleteById(userId);
-        } catch (NoSuchElementException exception) {
-            throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
-        }
+        Optional<User> optionalUser = userRepository.findById(userId);
+        if (optionalUser.isEmpty()) throw new ServiceException(REPOSITORY_ERROR__USER__ID_NOT_IN_REPO__ID);
+        userRepository.deleteById(userId);
     }
 }
